Move spectral feature computation to mlpnet
