+++ EEG2Video/GLMNet/modules/models_paper.py
@@ -16,6 +16,7 @@ import itertools
 import datetime
 import time
 import numpy as np
+from EEG_preprocessing.DE_PSD import DE_PSD
 
 import torch
 import torch.nn as nn
@@ -387,6 +388,15 @@ class mlpnet(nn.Module):
             nn.GELU(),
             nn.Linear(256, out_dim)
         )
+
+    @staticmethod
+    def compute_features(raw: np.ndarray, fs: int = 200, win_sec: float = 0.5) -> np.ndarray:
+        """Compute DE features from raw EEG."""
+        feats = np.zeros((raw.shape[0], raw.shape[1], 5), dtype=np.float32)
+        for i, seg in enumerate(raw):
+            de = DE_PSD(seg, fs, win_sec, which="de")
+            feats[i] = de
+        return feats
         
     def forward(self, x):               #input:(batch,C,5)
         out = self.net(x)
